requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
curl_cffi==0.6.2
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1